requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
lxml==4.9.1
numpy==1.23.1
oauthlib==3.2.0
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytrends==4.8.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.10.0
urllib3==1.26.11